requirements.txt
###############################################
##              requirements.txt             ##
###############################################

# ----------------- Build Tools -----------------

cython
setuptools>=68.0
wheel

# ----------------- Core API Stack -----------------

fastapi[standard]==0.115.3
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
minio==7.2.15
weaviate-client==3.26.7
pandas==2.2.2
numpy==1.26.4           # Latest stable release in the 1.x series for maximum compatibility
pydantic==1.10.14       # Required compatibility for SpaCy and FastAPI
pydub==0.25.1

# ---------------- Optional Audio & Video ----------------

librosa
ffmpeg-python
opencv-python-headless  # Use headless for server deployments
pillow
scenedetect
pytesseract

# ------------------- AI Stack (CPU) --------------------

torch==2.2.2
torchvision==0.17.2
# spacy==3.5.3
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
transformers>=4
keybert>=0.9

openai>=1.6
openai-whisper>=2024.09.30
git+https://github.com/openai/CLIP.git

# ---------------- LangChain Stack ----------------

langchain-core>=0.3.66
langchain-openai>=0.3.25
langchain-community>=0.0.34

# ------------------ NLP & Transformers ------------------

transformers==4.39.3
sentence-transformers==2.2.2
keybert==0.9.0
tokenizers==0.15.2
huggingface_hub==0.25.2
accelerate==0.27.2

# ------------------ Object Detection --------------------

ultralytics